+++ src/client/storage/index.js
@@ -3,20 +3,20 @@ import File, {FILE_DOWNLOAD_STATUS, FILE_UPLOAD_STATUS} from '../../db/models/fi
 import getDownloadQuery from './query';
 import {request} from 'graphql-request';
 import {
-    hashFn,
     merkle,
     delay,
     areScalarArraysEqual,
-    escape,
-    resolveHome
-} from '../../core/utils';
+    resolveHome,
+    statAsync
+} from '../../util';
 import {storage} from './storage';
 import {promises as fs} from 'fs';
 import path from 'path';
 import config from 'config';
 import logger from '../../core/log';
 import {uploadLoop, chunkValidatorLoop} from './uploader';
-import {statAsync} from '../../util';
+import {keccak256} from 'ethereumjs-util';
+import {encode} from 'html-entities';
 
 const log = logger.child({module: 'Storage'});
 
@@ -76,7 +76,7 @@ const getChunk = async (chunkId, encoding = 'utf8', useCache = true) => {
 
             const buf = Buffer.from(data);
 
-            const hash = hashFn(buf).toString('hex');
+            const hash = keccak256(buf).toString('hex');
             if (hash !== chunk.id) {
                 log.warn(
                     {chunkId, hash, query, buf: buf.toString()},
@@ -103,7 +103,7 @@ const getChunk = async (chunkId, encoding = 'utf8', useCache = true) => {
 };
 
 const uploadChunk = async data => {
-    const chunkId = hashFn(data).toString('hex');
+    const chunkId = keccak256(data).toString('hex');
     const chunk = await Chunk.findByIdOrCreate(chunkId);
 
     if (chunk.id !== chunkId) {
@@ -152,7 +152,7 @@ const uploadFile = async data => {
     const totalChunks = Math.ceil(buf.length / CHUNK_SIZE);
 
     if (totalChunks === 1) {
-        const fileId = hashFn(buf).toString('hex');
+        const fileId = keccak256(buf).toString('hex');
         log.debug({fileId}, 'File to be uploaded and consists only from 1 chunk');
 
         const filePath = path.join(filesDir, `file_${fileId}`);
@@ -200,10 +200,10 @@ const uploadFile = async data => {
         chunks.push(buf.slice(i * CHUNK_SIZE, (i + 1) * CHUNK_SIZE));
     }
 
-    const chunkHashes = chunks.map(chunk => hashFn(chunk).toString('hex'));
-    const merkleTree = merkle.merkle(
+    const chunkHashes = chunks.map(chunk => keccak256(chunk).toString('hex'));
+    const merkleTree = merkle(
         chunkHashes.map(x => Buffer.from(x, 'hex')),
-        hashFn
+        keccak256
     );
     const chunkInfoContents =
         CHUNKINFO_PROLOGUE +
@@ -217,7 +217,7 @@ const uploadFile = async data => {
     const chunkInfoBuffer = Buffer.from(chunkInfoContents, 'utf-8');
 
     // File id always matches it's index chunk id
-    const fileId = hashFn(chunkInfoBuffer).toString('hex');
+    const fileId = keccak256(chunkInfoBuffer).toString('hex');
 
     log.debug({fileId}, 'Successfully chunkified file');
     const filePath = path.join(filesDir, `file_${fileId}`);
@@ -277,16 +277,16 @@ const uploadDir = async dirPath => {
         const stat = await statAsync(dirPath);
         const isDir = stat.isDirectory();
         if (!isDir) {
-            throw new Error(`Path ${escape(dirPath)} is not a directory`);
+            throw new Error(`Path ${encode(dirPath)} is not a directory`);
         }
     } catch (e) {
         if (e.code === 'ENOENT') {
-            throw new Error(`Directory ${escape(dirPath)} does not exist`);
+            throw new Error(`Directory ${encode(dirPath)} does not exist`);
         }
         throw e;
     }
 
-    log.debug({dirPath: escape(dirPath)}, 'Uploading directory');
+    log.debug({dirPath: encode(dirPath)}, 'Uploading directory');
 
     const files = await fs.readdir(dirPath);
     const dirInfo = {
@@ -324,7 +324,7 @@ const uploadDir = async dirPath => {
 
     const id = await uploadFile(JSON.stringify(dirInfo));
 
-    log.debug({dirPath: escape(dirPath)}, 'Successfully uploaded directory');
+    log.debug({dirPath: encode(dirPath)}, 'Successfully uploaded directory');
 
     return id;
 };
@@ -382,11 +382,10 @@ const getFile = async (rawId, encoding = 'utf8', useCache = true) => {
         if (hash !== 'keccak256') {
             throw new Error('Bad hash type');
         }
-        const merkleReassembled = merkle
-            .merkle(
-                chunks.map(x => Buffer.from(x, 'hex')),
-                hashFn
-            )
+        const merkleReassembled = merkle(
+            chunks.map(x => Buffer.from(x, 'hex')),
+            keccak256
+        )
             .map(x => x.toString('hex'));
         if (!areScalarArraysEqual(merkleReassembled, merkleHash)) {
             throw new Error('Incorrect Merkle hash');
