+++ src/client/zweb/deployer/index.js
@@ -2,7 +2,7 @@ const path = require('path');
 const fs = require('fs');
 const logger = require('../../../core/log');
 const log = logger.child({module: 'Deployer'});
-const {compileContract, getImportsFactory} = require('../../../util/contract');
+const {compileContract, getImportsFactory} = require('../../../util');
 const {getNetworkPublicKey} = require('../../../wallet/keystore');
 const blockchain = require('../../../network/blockchain');
 const hre = require('hardhat');
