+++ src/util/index.ts
@@ -3,3 +3,7 @@ export * from './migrate';
 export * from './readFileByPath';
 export * from './makeSurePathExists';
 export * from './statAsync';
+export * from './delay';
+export * from './areScalarArraysEqual';
+export * from './resolveHome';
+export * from './merkle';
